requirements.txt
fastapi
uvicorn
python-dotenv
httpx
youtube-transcript-api
trafilatura
openai
jinja2
python-multipart